Load and Save
Here’s a more specialized translation of your text, suitable for the GitHub Description section:

Loading and saving the initial data posed several challenges, including:

Large data volume: The product dataset was approximately 1.82 GB, while the customer search data was around 1.53 GB. Converting them to CSV format resulted in file sizes of 659 MB and 1.27 GB, respectively.

Data loading time: Due to the significant size and the fact that the datasets were in JSONL format (with each data record stored in a separate line), I utilized a Python technique called "chunking." This method allowed for incremental data loading by processing a specified number of lines at a time. The "chunksize" parameter was adjusted based on the available hardware to optimize performance.

Time-consuming product data processing: To mitigate this, I created a 1/10 sample of the product data for initial testing and saved it for future use. We might also need to apply this sampling to the customer search data, though at this point, it doesn't seem necessary.